pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lib_streamlit_chat_with_fileupload"
version = "0.1.0"
description = "Streamlit custom component that offers chat input box for messages - with files upload widget"
authors = [
    { name = "Stellars Henson \u003ckonrad.jelen@gmail.com\u003e" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = "~=3.12.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "boto3",
    "python-dotenv",
    "streamlit",
]

[project.optional-dependencies]
dev = [
    "build",
    "ipykernel",
    "ipython",
    "nbdime",
    "pip",
    "pytest",
    "pytest-cov",
    "toml",
    "ruff",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["lib_streamlit_chat_with_fileupload*"]
exclude = ["tests*"]

[tool.ruff]
line-length = 99
src = ["lib_streamlit_chat_with_fileupload"]
include = ["pyproject.toml", "lib_streamlit_chat_with_fileupload/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["lib_streamlit_chat_with_fileupload"]
force-sort-within-sections = true
